MB-45071: Pass DcpProducer& to ActiveStream::notifySeqnoAvailable()
Remove notifyStreamAvailable() from the Stream parent class (unused
for PassiveStream) and change the seqno parameter (also unused) to
a DcpProducer& that is passed in from the caller (the DcpProducer).
Doing this avoids the need to promote the producerPtr weak_ptr when
we notify the DCP connection of a new seqno.

Change-Id: I4355d6cf1248424009201d946d7dd823a8fdf871
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/150249
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>