MB-45044: Regression test for exception thrown via std::thread
Due to GCC Bug 55917 (present gcc-7.3), uncaught exceptions thrown
from a std::thread do not report the backtrace of where the exception
was thrown - only where it was caught in std::thread machinary.

Add a regression test for the issue, which is addressed by our upgrade
to a newer GCC (10.2, but 8 would in theory be sufficient).

Change-Id: I5965034ff9a2c9210efa7ad507be92eb059d79c7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/150473
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>