MB-45505: Add quorum_history_id to bucket config
To ensure there is no window where KV does not know the current
history-id, update the configuration so that ns_server can give
KV the current history-id when the bucket is created.

If collections are in the 'default' state (always the case
for ephemeral) we can default initialise our view of the id
from the given value.

The result is there no window where KV is in a state of
"unknown history-ID" and KV does not have to consider all
of the consequence of what may happen when in this state.

Change-Id: I431bb8bea4ca627ed78481fb593e609fd538314e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/150673
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>