MB-43818: Fix breakpad_test GDB checks on Linux
As part of patch "MB-43818: Allow breakpad_test to run for macOS"
(e0a6dbb4d902), breakpad_test was inadvertently broken on platforms
supporting breakpad - the new '--breakpad' argument was not passed
meaning the breakpad-specific checks were not run.

Fix the argument passing in CMake to breakpad_test.py, so Breakpad and
GDB checks are now run on Linux.

Change-Id: Ifa497d88fd3963895f38b8aecb0d5d5121233c71
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/150655
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>