MB-36227: Make VBReadyQueue benchmark follow access pattern
Make the MPSCRandom VBReadyQueue benchmark follow the typical access
pattern of the VBReadyQueue (to pop until a pop returns false).
Currently with this variant of the VBReadyQueue the benchmark runs
without getting hung. With the lock free variant of the VBReadyQueue
though there is a "race condition" where the producer may not
notify the consumer if a concurrent pop is happening. This is dealt
with safely by the consumer popping until the first pop returns false.
The benchmarks needs to reflect this as it otherwise hangs waiting for
a notification that does not come.

Change-Id: Ib0e4c9d4abfb836ec1457c0e04be1f7b997f0c39
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/150972
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>