MB-45459: Pass old item to persistenceFailureStatOvercounts
Currently we pass the old queue time to the
persistenceFailureStatOvercounts stats (stat changes applied if a
flush fails) which is required as the queue time may change for the
next item flushed. The same issue exists for dirtyQueuePendingWrites
which may change if the item changes size. Pass the entire old item
to the persistenceFailureStatOvercounts instead.

Change-Id: I9caa796522b876a5b6d96685832951aac0d8579c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/151033
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>