MB-45662: Add ability to iterate though buckets to mctimings
Add the ability to mctimings to display timing histograms on a
per-bucket level. Currently there's no way to do this unless the user
explicitly specifies all the buckets on a node. Otherwise the user can
receive the histograms aggregated across all buckets by default or when
the '/all/' bucket is specified.

To add this functionality use MemcachedConnection::listBuckets() to get
hold of all the buckets and then iterate though each bucket displaying
histogram data for ops in the same way we would if just one bucket was
selected. Though also some printing to display the bucket that we're
currently displaying. This functionality is enabled by a new cli arg -a
or --all-buckets which takes no argument.

Change-Id: I6b179e5e4e07cf2292ca6046ff866417afd82d35
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/151167
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>