Fix mctiming's short form file argument -f
Fix mctiming's short form file argument '-f' currently this doesn't work
as it needs to be specified as part of 'optstring' for getops_long(). To
fix this add 'f:' to the end of the string which tells getops_long() to
expect '-f' arg, which must have an argument.

Also take this as an opportunity to move long_options[] to a
std::vector<> to remove clang-tidy warning.

Change-Id: I893ff20888a7d14a26ebad166653cef30716e71a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/151172
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>