MB-45378: Remove uses of old-style FOLLY_xxx variables
In preparation for ensuring all targets which link to folly are
statically linked, remove uses of old-style FOLLY_INCLUDES /
FOLLY_LIBRARIES variables; preferring the "modern CMake" targets.

This also required some headers to be shuffled around where not
required in the .h file.

See also the patches to couchstore, platform and tlm with same
Change-Id.

Change-Id: I9e0f9474e93ceddc92d324406308688b545e5c4a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/151233
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>