Refactor unit test for bucket name and type validation tests
There isn't much value in testing the bucket name validation
tests by going over the wire when it can be run in a tiny
isolated unit test instead.

And add unit tests for validating all of the legal modules
which may be provided with the create bucket command.

Move the name validation to the package validator (we don't
need to start a new thread to look if the key is too long
or contains invalid characters)

Change-Id: I627e00ceff4317f6de3a1fe43e96519371bb2edb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/151525
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>