MB-45523: Fix race in takeover blocking when buffer log full
There is a race condition in the blocking of front end ops during
a takeover where if the buffer log is full we unblock operations
without transitioning from takeoverSend(pending) to
takeoverWait(pending). This could cause us to indefinitely hang
during a rebalance if the rate of operations is high enough.

Correct this race condition by moving the unblocking of ops to
after we transition the stream state.

Unblocking operations is of questionable value here and if we have
blocked operations then unblocking them again will only serve to
make takeover longer. That issue should be addressed, but is
orthogonal to this issue.

Change-Id: I3b04165093170756628c9370d827c27f728865d9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/151752
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>