MB-45757: Make CM::removeCursor resilient to invalid cursors
In CheckpointManager::removeCursor we check that the cursor being
removed is valid. If we find that we are removing an invalid cursor
then we throw and crash.

But, the current CheckpointCursor concurrency model allows legal (and
safe) conditions where we may try to invalidate/remove a DCP cursor
twice, so the fix is to just do nothing and return the call in that
case.

Change-Id: I95678ba459d9ce0058fe46fe462da60f445b5f32
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/151845
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>