MB-45892: Clients should be able to HELO collections on any 7.0 node.
The HELO executor denies 'collections' during mixed mode, partially
because the code was implemented when collections were a DP feature.

During mixed-mode running, collections are enabled and disabled
depending on your perspective.

The cluster effectively disabled collections, the user cannot make
changes to the collection state. Additionally DCP for replication runs
without collections.

However an individual 7.0 node, collections are enabled (the default
collection exists and should be 100% operational as a collection). A
collection aware client should be able to use the collection protocol
changes with no issues in mixed-mode operation, the default collection
is all that exists and they should be able to explicitly reference that
collection.

The Settings::isCollectionsEnabled changes from false/true when the
cluster is no longer mixed and collections can be created/dropped, with
that it is not needed in the HELO path.

Change-Id: Ib39f4b4ec529bc68d8a3eca4ae3e0c0635bffc26
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152080
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>