MB-45755: Don't start flushers until VBMap populated
During warmup we flush the vBucket state to disk to ensure that
we can write to disk. If we cannot write to disk then we do not
want to enable traffic. A race condition currently exists during
warmup where a warmup thread and a flusher can attempt to flush
different vBuckets on the same shard. This can cause us to flush
items or state against the wrong vBucket or crash if we corrupt a
variable in memory as the KVStore (per shard) uses member variables
during flushing.

The simplest solution to this problem is to simply not schedule
the flusher until the PopulateVBucketMap phase of warmup has run
for the given shard. This ensures that the flusher will not attempt
to flush to the shard until after warmup is done with it.

Change-Id: Ic0527519b77b63380cc769a61685548a25676c6a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152102
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>