MB-45756: Expand logging when don't find a pending item
Add extra logging for when we don't find a pending items in the
HashTable while trying to commit a item. Currently we just provide
information about the prepare/commit seqnos and the key. Expand this to
provide a better context of what the current snapshot state and DM
seqnos are.

Change-Id: Ie8eb2fadcd2d03f4674971760705afb723274ed7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152179
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>