MB-45942 magma-kvstore: Fix expectation in RollbackNoValidCheckpoint test
The magma storage engine maintains checkpoints equal to configured
number of maxCheckpoints or configured time intervals for rollback.
There is also a configured max overhead limit based on disk size.
In addition, it may internally maintain an additional checkpoint for
recovery. Hence, the expected number of checkpoints is equal to
maxCheckpoints or maxCheckpoints+1 (If it generates an internal recovery
checkpoint). Earlier a bug in magma checkpoint overhead calculation resulted
in creating checkpoints aggressively and retaining checkpoints maxCheckpoints
and the test was passing consistently. The change corrects the expected value.

Change-Id: Ib1088c5d4849e0ba7d68b9bd10f56a45ec5d634a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152189
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>