MB-45756: Expand PDM logic exception exception message
If we receive a commit out order add extra information
to the log message to tell us if there is a prepare that is being
tracked for the commits key and if so what seqno it has. Plus log
information around the current HPS, HCS and last seen snapshot end.

When creating a exception message when there are no trackedWrites use
fmt::format() over a std::stringstream as its easier to read and grep in
the code base.

Change-Id: I7f6c0642bf38e4c7ca0171d5e74c76f7be0ad778
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152359
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>