MB-42272: Track dropped collection stats for magma
Magma can't track collection stats in the same way that couchstore
can when it comes to erasing dropped collections. This is because
magma may visit stale versions of keys during compaction and we can't
check if the item was the newest at the point at which we started
the compaction. As such, when a collection is dropped for a magma
bucket we just decrement the item count by the amount of items
recorded in the collection stats doc. Currently as only one doc exists
to track the stats of a collection we cannot track stats correctly
if a collection is dropped and recreated before the erasure
(compaction) runs. This is because new collections will
overwrite the existing collection stats doc.

Track a stats doc for dropped collections that are pending erasure
for magma buckets which is ammended if new generations of a
collection are dropped with the total values of items that should be
erased.

Change-Id: Ie9d3d6a015033dfaff0df6e9f89fe241607f32bc
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/149040
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>