Remove CheckpointTest::basic_chk_test
This was the very first (generic) test added when the Checkpoint and
CheckpointManager classes were introduced.
That is the only test that uses the queue_op::flush enum entry that we
aim to remove. Given that now we have thousands of tests that cover the
checkpoint logic in many specific areas then it seems reasonable to
remove the basic_chk_test and proceed with getting rid of the legacy
queue_op::flush.

Change-Id: I390df7380bc70eeb7a09d3f79cb80654a2918938
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152347
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>