MB-45756: Log the seqno of the first & last prepare found
During EPVBucket::loadOutstandingPrepares(), we log information about
the high completed seqno and prepare seqnos when loading prepares from
disk for the DM. We also log how many prepares we've found on disk and
the high persisted prepare seqno. Though this is useful it doesn't give
us an idea of where in seqno list the first non completed prepare is.
So log the seqno of the first item in outstandingPrepares (which is
ascending sorted). We also log the highest prepared seqno found as this
can differ from the PPS we're only received a partial disk snapshot
which contains seqnos that haven been logically completed.

Change-Id: Ie9c0b511fe716c9e6063f1815234dc05df3759ec
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152357
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>