MB-39745: RO KVstore should return const RW KVStore
Currently for couchstore we use the RO KVStore for read use cases.
The RO KVStore uses a significant amount of memory unnecessarily
as all usages of the RO KVStore can be performed by the RW KVStore.

Change the function returning the RO KVStore to return s const RW
KVStore so that in a future patch we can remove the RO KVStore
entirely.

Change-Id: I44176b94e5b1da00d58e71c94a53de30a808c287
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152890
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>