MB-39745: Don't create RO KVStore
This patch reduces memory usage by not creating the RO KVStore now
that we don't use it anywhere.

Interestingly, one of the ItemPagerTests now starts to fail as it
doesn't create enough items on the replica. The reason for this is
because when we call increaseQuota(n) we add the value n to the
current memory usage and se the quota to that rather than increase
the quota by n. To account for the 200KB less memory used in this
test we need to increase the quota by a further 200KB for the test
to keep working as it previously did.

Change-Id: Ifef7aa17f16633bb7fbcf8163f366bc6fafd1f90
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/152891
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>