MB-45756: Refactor snapshotStats methods to use nlohmann::json
Refactor snapshotStats to use nlohmann::json as they use a
std::map<std::string, std::string> which needs to be converted to a json
format before being persisted to disk. nlohmann::json can do the
conversion from C++ object to a string for us and provides a better
interface for read/manipulating the json data.

This will also help with follow up patches for MB-45756.

Change-Id: Idaa253a91bcbab894eff616e0d5650ee96506879
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153027
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>