MB-46173: isSuccessor failing when new manifest appears equal.
The manifest isSuccessor code is too strict and failing a valid input.

It's possible for the following manifest state changes:

* A {c1, c2}
* B {c1, c2, c3} # create c3
* C {c1, c2} # drop c3

Here if KV were to only observe A then C, we would fail the update.
This situation could occur if a node was 'offline' briefly or the
B->C change occurred quickly and ns_server only pushed out C.

The Manifest validation is complaining that C is invalid, the uid
increased but made no change to our collection state, this is
an incorrect assumption and is now removed.

Change-Id: I5666aa1fec8d53d21ee939ecdf4a1e4e2e91082b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153118
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>