MB-46046: [1/7] Make HdrHistogram iters follow typical C++ iter semantics
Begin updating HdrHistogram::Iterator to meet more requirements of
`LegacyInputIterator`.

With this, histograms can be iterated in the typical C++ style

 for (auto iter = foo(); iter != end(); iter++)

Or, for the default iteration mode:

 for (const auto& bucket: histogram)

This patch does not update all users of the iterators, rather
`Iterator::getNextValueAndCount()` and friends remain as shims for now,
implemented using `operator++`. These will be removed in a future
patch.

Change-Id: I5f07e3673ea0c93492c02b77622bbb2851ba839c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153279
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>