MB-46216: Introduce EP_LOG_<level>_RAW macros
In preparation for making the existing EP_LOG_<level> macros check the
format string at compile-time, we need to migrate any log calls which
_don't_ use a format string (i.e. just print an unformatted string
literal) away from EP_LOG_<level> to a new macro set -
EP_LOG<level>_RAW.

This is due to limitations in fmtlib's compile-time format string
checking - essentially a sting like "foobar" isn't considered a valid
format string (it doesn't contain any '{}' placeholders). As such, we
cannot unconditioanlly apply the compile-time format check to the
first argument to be logged; only if there's more than 1 argument (and
hence first should be a format string). It might be possible (with
clever preprocessor usage) to conditionally apply the macro at
compile-time based on argument count, but this is non-trvial;
particulary to get working on all our supported compilers.

Add the following macros to log a type implicitly-convertable to const
char*, or which supports operator<<:

* EP_LOG_TRACE_RAW
* EP_LOG_DEBUG_RAW
* EP_LOG_INFO_RAW
* EP_LOG_WARN_RAW
* EP_LOG_ERR_RAW
* EP_LOG_CRITICAL_RAW

Change-Id: I4ae47eacf1fe435aca24096a845c9ba45de8c122
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154686
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>