MB-46216: Add support for using external fmtlib via spdlog
Allow kv_engine to find correct fmtlib includes when using an external
(non-bundled) fmtlib with spdlog. Previously we just added the include
directory of spdlog to the include search path, however that fails to
set the correct preprocessor definition (SPDLOG_FMT_EXTERNAL) which
ensures the spdlog headers use an external fmtlib.

Fix by removing that path, and instead link to the 'spdlog' interface
library which sets the correct #define (and adds the correct search
path to fmtlib).

Change-Id: I912fda4236ae8cc3a3702bad7ab2ac7ecc7e1990
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153271
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>