MB-46203: Improve logging when persistence wait times out
The associated MB is not triggered by a regression, but code that
could be improved. As part of that improvement, and any other
rebalance failures triggered by this code capturing the values
used in the "failure" will be useful.

Logging now captures the time spent flushing and the value of the
limit (which changes at runtime).

Change-Id: If75bc9615d762b6a96270e4d8d019f30fcf5198a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153234
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>