MB-46046: [2/7] Introduce cb::iterator_range
Provides a simple container for two iterators (or an iterator and a
sentinel) representing a range.

Exposes begin() and end(), and may be used to return a value useable in
a range-based for loop, e.g.,

 for (const auto& foo: bar.some_view()) {...}

Useful if a class provides multiple, logically different ways of
iterating over its contents.

Change-Id: Ia2dcde54199c0c3d58ca9a8a4419a1375eb97b71
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153280
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>