MB-46046: [5/7] Support iterating all possible buckets for HdrHistogram
Normally HdrHistogram iterators stop once the iterated buckets cover
all recorded values. This is reasonable, as any further buckets will be
empty (or unmeaningful, for Percentile or Recorded iterators).

For Log and Linear iterators however, the buckets boundaries are
predetermined. Some callers may wish to iterate over _all_ the buckets
the histogram will ever contain, to ensure they see a stable set of
buckets over time.

To support this,

 histogram.linearViewRepresentable()

and

 histogram.logViewRepresentable()

have been added to complement linearView() and logView() respectively.

These methods return iterator ranges which will extend to the max
representable value of the histogram.

The set of buckets seen will not change as more values are added, and
any buckets past the highest recorded value will simply be empty.

Change-Id: Ie3cc54841fa9997bd5d830eb5e02736cb231652f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153332
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>