MB-46017: Refactor: Use ProcessMonitor from platform
The functionality in ParentMonitor was moved into a more generic
library function ProcessMonitor and moved to platform so that
we can reuse the functionality elsewhere (to spawn and monitor
memcached process in testapp and cluster_test).

Extend cluster_test to detect when the underlying memcached
process crash, and if it takes more than a minute to
run (and dump the last 8k of the log file if it does). To
try to give more information on why the program hangs every
now and then on windows.

Change-Id: I48e8b51d852099ccc82b09fbbd498e1b4b1fd0a0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154054
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>