MB-46477 [BP 6.6.3]: Subdoc: Only log one DOC_MODIFY event
When performing a subdoc single or multi-mutation; two DOCUMENT_MODIFY
audit events are logged. One is (correctly) logged by the underlying
bucket_store() function; but an additional duplicate event is logged
when forming the response to the client.

This is incorrect; we only need to log as part of the response for
DOCUMENT_READ events (as these are not logged internally).

(Backport of commit bb8c5ea to mad-hatter branch)

Change-Id: Ibf898c2b3d3db9d56dabbd4c9d368c09113c9a97
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153555
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>