MB-46310: Only persist ep_force_shutdown from destruction thread
Currently the StatSnap task can persist the ep_force_shutdown stat
with a value of "false" during a clean shutdown if it runs after
we set the shutdown stats. If a shutdown were to run slowly due to
having to flush a large number of items to disk then the shutdown
could potentially crash after we have persisted this stat. Correct
this by only setting the shutdown stats from the bucket destruction
thread.

Change-Id: I9744ff6c45d7de88c8de5968dacf769c0c564a44
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153744
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>