MB-46203: Simplify 'wait_for_persistence' time-out
Remove all code which tried to be 'smart' and dynamically adjust
the timeout based on actual performance - it doesn't work well if
the workload or system load shifts from 'fast' to 'slow'.

For now just set the time-out at the old 'max', which is 30 seconds.

Change-Id: I78481c801581719273ae6a4bfc793587ede7140f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/153812
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>