MB-43299: Remove IndexEntry::mutationId
Back to some commits ago, IndexEntry::mutationId was used at
de-duplication for getting the seqno of items pointed by cursors. Usage
removed in http://review.couchbase.org/c/kv_engine/+/148556.

After that, the mutationId is now used at de-duplication for only for
detecting a SyncWrite that has been expelled. Ie, we set it to 0 when
the SW is expelled, and than at dedup we identify an expelled SW by
mutationId==0 and we enforce the proper dedup logic accordingly.

In this patch we leverage the existing IndexEntry::position logic to
cover the "expel SW" case. So mutationId becomes redundant and can be
removed.

IndexEntry::position is already use at expel/dedup for marking an
"expelled normal mutation". That by setting the position to
Checkpoint::toWrite::end().

The strategy in this patch is to use the Checkpoint::toWrite::begin()
special position to invalidate the IndexEntry for SW. Note that (same as
C::end()), the C::begin() always exists in Checkpoint as the empty item.

With this patch we push down the mem-usage of a single IndexEntry from
16 bytes to 8 bytes.
The mem-usage gain is not directly visible in our microbenchmarks
(probably due to mem-alignment at allocations), but that is something
compiler-dependent, so we may see some effect on some platforms.

Change-Id: Ic7e31bbde6efb0b568c73dc2685dd52e5b2f8a71
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154052
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>