MB-45026: Don't expel items pointed by cursors
Currently there is an inconsistency in the behaviour, as an item pointed
by a cursor may be expelled or not depending on whether it is at
high-seqno.

Here we make the behaviour consistent by always preserving items
referenced by cursor. That also prevents any issue related to the
possibility of having invalid cursors around.

Change-Id: I4c5b10d89bba88da7894c780a2490148214b170f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154152
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>