MB-46517: crash_engine: Stop compiler optimizations from skipping crash
With GCC-10.2 using build type RelWithDebInfo, the compiler is being
too clever and optimizing away the illegal memory access used to
trigger a crash. As a result the test doesn't crash and hangs,
eventually timing out when entire job times out.

Issue seems to be related to it recognising the global dummy value is
only ever accessed from that function, and we don't provide a default
value for it.

Fix by marking the crash function as noinline (to avoid unrolling that
way), and the memory address we dereference volatile.

Also add a timeout to breakpad-test so it doesn't hang forever in
future.

Change-Id: I9d15f63fb9c509cd3ae07e087e0ddf86e9ef7c1b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154329
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>