MB-46482: Producer to account for namespace of keys for commit/abort
The producer was always accounting in the buffer log for key size
with the namespace prefix which in mad-hatter is not transmitted
to the consumer (for DCP commit and abort). This leads to the consumer
acking for the key size without the prefix, leading to a DCP pause.

When creating a producer abort or commit message, use the 'collection'
mode of the stream to account for the correct key size.

Change-Id: I4a364463f94eec591ecf6a875f6b83ebfe3ee03d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154997
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>