MB-33758: Don't expect snapshot marker count during takeover
We shouldn't expect a given number of snapshot markers in the
dcp stream takeover test as checkpoints may be cut differently
run to run. Normally we could just stream from disk but for a
takeover stream that is a different behaviour. Instead, don't
add any expection to the number of snapshot markers.

Change-Id: Ic116fdf2c1eda2b069fbdf23d23d7c2e73b046fd
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154473
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>