Remove the deprecated LockHolder
According to engines/ep/README.md the LockHolder was deprecated.
The motivation is that atomic.h included locks.h and atomic.h
is needed by the executor we want to make usable from the core.

Remove LockHolder from locks.h and rename the file to
multi_lock_holder.h to better reflect its content and only include
the file where it is needed.

Change-Id: I9533390c140a513c49a279ca54ec6977336b1c93
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154513
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>