MB-46592: Catch exceptions that may occur inside logExecutionException
The MB identified that calls to "toJSON" whilst handling an "execute"
exception have been seen to throw a further std::exception, this is
not caught and leads to memcached being terminated.

The fix wraps the code that is building a detailed exception in
a try/catch which then leads to a simpler attempt of logging
the error.

Change-Id: I0ef771ab81fc057b7a0e2ac1215f39ad8b66fc2e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154543
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>