MB-46628: [BP] Consumer must be able to call CommitSyncWrite::getMessageSize
An update to commit/abort handling added an exception to both
CommitSyncWriteConsumer and AbortSyncWriteConsumer, this was
incorrect and meant that when DCP messages are buffered, we
hit the exception.

This commit tweaks the CommitSyncWrite hierarchy and allows
getMessageSize to be called for both producer/consumer.

Test cases added in this commit noted that DcpPrepare has
an incorrect buffer size, now tracked as MB-46634. This
commit adds very basic documentation for dcp prepare to
make this issue more evident.

Change-Id: Ic504de6103f8bdbf9f9f258d6bcbf01b55a7e408
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/154998
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>