MB-39620: Add support for ifconfig commands [2.0/3]
This patch contains the implementation for the ifconfig
commands (and a unit test suite which tests them).

It is currently not possible to fully use the new commands
as it is in "conflict" with the normal config reload (it'll
give you undefined behavior as it might drop and open ports
etc).

Change-Id: I56eea6cbcb2e0d4b073f4247c29d7308dc40d67b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129645
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>